requirements.txt
g4f==0.5.2.2
python-dotenv==1.1.0
aiogram==3.20.0.post0
